chore(popup): centralize logic in a single class PopupApp
