+++ chrome/styles/base.css
@@ -1,3 +1,19 @@
+@font-face {
+  font-family: "Rubik Moonrocks";
+  src: url("../assets/fonts/RubikMoonrocks-Regular.woff2") format("woff2");
+  font-weight: 400;
+  font-style: normal;
+  font-display: swap;
+}
+
+@font-face {
+  font-family: "Comfortaa";
+  src: url("../assets/fonts/Comfortaa-VariableFont_wght.woff2") format("woff2");
+  font-weight: 400;
+  font-style: normal;
+  font-display: swap;
+}
+
 *,
 *::before,
 *::after {
@@ -6,6 +22,29 @@
 
 body {
   margin: 0;
+}
+
+:root {
+  --font-family: "Rubik Moonrocks", sans-serif;
+  --font-family-second: "Comfortaa", sans-serif;
+
+  --color-grey: #898989;
+  --color-white-80: rgba(237, 237, 237, 0.8);
+  --color-white: rgb(237, 237, 237);
+  --background-dark: #1a1a2e;
+  --background-accent: #0f5cad;
+  --light-blue: #2e85e1;
+  --color-green-element: #4caf50;
+  --color-red-element: #f44336;
+
+  --container-overlay-blue: rgba(15, 91, 173, 0.4);
+  --container-overlay-dark: rgba(26, 26, 46, 0.8);
+  --shadow-glow-light: rgba(250, 250, 250, 0.05);
+
+  --dark-green-button: rgb(35, 113, 26);
+
+  --dark-green-default: #299e1f;
+  --lite-green-active: #97e851;
 }
 
 h1,
