--- chrome/styles/fonts.css
@@ -1,15 +1,0 @@
-@font-face {
-  font-family: "Rubik Moonrocks";
-  src: url("../assets/fonts/RubikMoonrocks-Regular.woff2") format("woff2");
-  font-weight: 400;
-  font-style: normal;
-  font-display: swap;
-}
-
-@font-face {
-  font-family: "Comfortaa";
-  src: url("../assets/fonts/Comfortaa-VariableFont_wght.woff2") format("woff2");
-  font-weight: 400;
-  font-style: normal;
-  font-display: swap;
-}
