+++ chrome/styles/styles.css
@@ -1,4 +1,2 @@
-@import url("variables.css");
 @import url("base.css");
-@import url("fonts.css");
 @import url("animation.css");
