--- chrome/styles/variables.css
@@ -1,22 +1,0 @@
-:root {
-  --font-family: "Rubik Moonrocks", sans-serif;
-  --font-family-second: "Comfortaa", sans-serif;
-
-  --color-grey: #898989;
-  --color-white-80: rgba(237, 237, 237, 0.8);
-  --color-white: rgb(237, 237, 237);
-  --background-dark: #1a1a2e;
-  --background-accent: #0f5cad;
-  --light-blue: #2e85e1;
-  --color-green-element: #4caf50;
-  --color-red-element: #f44336;
-
-  --container-overlay-blue: rgba(15, 91, 173, 0.4);
-  --container-overlay-dark: rgba(26, 26, 46, 0.8);
-  --shadow-glow-light: rgba(250, 250, 250, 0.05);
-
-  --dark-green-button: rgb(35, 113, 26);
-
-  --dark-green-default: #299e1f;
-  --lite-green-active: #97e851;
-}
